refactor(quality): extract CustomDictionaryDetectionRunner from AnonymizationService
The custom-dictionary detection pass was five cohesive private methods on an
already-oversized service (TooManyMethods 29). Moved verbatim into its own
runner; AnonymizationService now delegates one call.
