+++ tests/unit/Service/AnonymizationLinkServiceTest.php
@@ -179,8 +179,10 @@ class AnonymizationLinkServiceTest extends TestCase
             pdfConversion: $this->createMock(originalClassName: PdfConversionService::class),
             emlAssembly: $this->createMock(originalClassName: EmlPdfAssemblyService::class),
             customDictionary: $this->createMock(originalClassName: CustomDictionaryService::class),
-            customDictionaryMatch: new CustomDictionaryMatchService(),
-            confidentialityLabel: $this->createMock(originalClassName: \OCA\DocuDesk\Service\ConfidentialityLabelService::class)
+            confidentialityLabel: $this->createMock(originalClassName: \OCA\DocuDesk\Service\ConfidentialityLabelService::class),
+            customDictionaryDetection: $this->createMock(
+                originalClassName: \OCA\DocuDesk\Service\CustomDictionaryDetectionRunner::class
+            )
         );
 
     }//end buildService()
