+++ tests/unit/Service/AnonymizationServiceProhibitionTest.php
@@ -127,8 +127,10 @@ class AnonymizationServiceProhibitionTest extends TestCase
             pdfConversion: $this->createMock(originalClassName: \OCA\DocuDesk\Service\PdfConversionService::class),
             emlAssembly: $this->createMock(originalClassName: \OCA\DocuDesk\Service\EmlPdfAssemblyService::class),
             customDictionary: $this->makeNoOpCustomDictionaryService(),
-            customDictionaryMatch: new CustomDictionaryMatchService(),
-            confidentialityLabel: $this->createMock(originalClassName: \OCA\DocuDesk\Service\ConfidentialityLabelService::class)
+            confidentialityLabel: $this->createMock(originalClassName: \OCA\DocuDesk\Service\ConfidentialityLabelService::class),
+            customDictionaryDetection: $this->createMock(
+                originalClassName: \OCA\DocuDesk\Service\CustomDictionaryDetectionRunner::class
+            )
         );
 
     }//end makeService()
