+++ tests/unit/Service/AnonymizationServicePublicationClearanceTest.php
@@ -90,8 +90,10 @@ class AnonymizationServicePublicationClearanceTest extends TestCase
             pdfConversion: $this->createMock(originalClassName: \OCA\DocuDesk\Service\PdfConversionService::class),
             emlAssembly: $this->createMock(originalClassName: \OCA\DocuDesk\Service\EmlPdfAssemblyService::class),
             customDictionary: $this->createMock(originalClassName: CustomDictionaryService::class),
-            customDictionaryMatch: new CustomDictionaryMatchService(),
-            confidentialityLabel: $this->createMock(originalClassName: \OCA\DocuDesk\Service\ConfidentialityLabelService::class)
+            confidentialityLabel: $this->createMock(originalClassName: \OCA\DocuDesk\Service\ConfidentialityLabelService::class),
+            customDictionaryDetection: $this->createMock(
+                originalClassName: \OCA\DocuDesk\Service\CustomDictionaryDetectionRunner::class
+            )
         );
 
     }//end buildService()
